Backend/requirements.txt
# Python environment variable management
python-dotenv  

# Google GenAI API client
google-genai

# Flask web framework and related utilities
flask 
flask-cors  

# Utilities for running processes concurrently
concurrently 

# Data manipulation and machine learning
pandas 
sklearn 
numpy 

# MongoDB client for Python
pymongo 


npm install uuid
npm install framer-motion
python -m pip install chardet
